modularize auth.ts and create userController to handle user specific actions
