+++ backend/src/routes/userRoutes.ts
@@ -2,11 +2,10 @@ import { Router } from 'express';
 import {
     register,
     login,
-    protect,
-    getProfile,
-    updateProfile,
     verifyEmail
 } from '../controllers/auth.js';
+import { protect } from '../middleware/auth.js';
+import {getProfile, updateProfile,} from '../controllers/userController.js';
 
 const router = Router();
 
@@ -15,4 +14,5 @@ router.post('/login', login);
 router.get('/profile', protect, getProfile); 
 router.put('/profile', protect, updateProfile);
 router.post('/verifyEmail', verifyEmail);
+
 export default router;
